model changes

+++ pkg/model/model_package.go
@@ -9,7 +9,7 @@ type Package struct {
 	Description  string `json:"description"`
 	ConsultationService     *ConsultationService `gorm:"foreignkey:consultationserviceID" json:"consultationservice"`
 	ConsultationServiceID int64
-	Duration int64  `json:"duration"`
+	Duration float64 `json:"duration"`
 }
 
 func (Package) TableName() string {
